pyproject.toml
[project]
name = "musicbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.20.0.post0",
    "langchain>=0.3.25",
    "langchain-openai>=0.3.18",
    "python-dotenv>=1.1.0",
]
